+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 CHANGELOG
 =========
 
+2.3.0
+-----
+
+* Added LtiSubmissionReviewRequest message launch flow (platform and tool sides)
+* Added LtiEndAssessment message launch flow (platform and tool sides)
+* Added visual indicator for required form fields
+* Updated documentation
+
 2.2.0
 -----